backend/requirements.txt
flask
facenet-pytorch
torch
torchvision
scikit-learn
pillow
werkzeug
flask-cors
python-dotenv-1.0.1